refactor: move around types and fns to correct modules
removed some unecessary modules
