--- bevy_midix/src/synth/sink/inner.rs
@@ -1,9 +1,0 @@
-use midix::prelude::ChannelVoiceMessage;
-
-use crate::song::SongId;
-
-pub struct InnerCommand {
-    pub time_to_send: u64,
-    pub parent: Option<SongId>,
-    pub command: ChannelVoiceMessage,
-}
