+++ bevy_midix/src/synth/sink/mod.rs
@@ -4,9 +4,7 @@ This Sink will send events to another thread that will constantly poll/flush com
 */
 
 mod commands;
-pub use commands::*;
+pub(crate) use commands::*;
 
 mod task;
 pub(crate) use task::*;
-
-mod inner;
